+++ src/modules/icons.js
@@ -25,7 +25,8 @@ import {
   ChartColumn,
   ChevronLeft,
   ChevronRight,
-  ChevronsRightLeft
+  ChevronsRightLeft,
+  ArrowUpDown
 } from 'lucide';
 
 // Map of all icons used in the app (PascalCase keys for createIcons)
@@ -51,7 +52,8 @@ const icons = {
   ChartColumn,
   ChevronLeft,
   ChevronRight,
-  ChevronsRightLeft
+  ChevronsRightLeft,
+  ArrowUpDown
 };
 
 /**